Rename file UseInstalledPlugins to UseInstalledPluginsWithoutChannels.
