+++ plugins/woocommerce-admin/client/marketing/hooks/index.ts
@@ -1,4 +1,4 @@
-export { useInstalledPluginsWithoutChannels } from './useInstalledExtensions';
+export { useInstalledPluginsWithoutChannels } from './UseInstalledPluginsWithoutChannels';
 export { useRegisteredChannels } from './useRegisteredChannels';
 export { useRecommendedChannels } from './useRecommendedChannels';
 export { useCampaignTypes } from './useCampaignTypes';
